Adding to Seating Chart
Added the actual seating chart where each button is a seat and it is clickable. Still need to find the flow of pages from Showtimes -> checking out. What do we want the flow to be? Right now it's Showtimes > Purchase Tickets > Seating chart. This can be changed however we need it to be.