Added active movies MovieCard and MovieDetails on the Home and Schowings page
Added the active movies as of today in the redux store state for ease of access.

This is passed to the home page and showings page so the movie data can be easily used.

Moved the MovieCard and MovieDetails into a sigle component, and created the modal as its own component so it can be referenced directly by pages.